requirements.txt
playwright==1.44.0
google-generativeai==0.5.0
chromadb==0.4.24
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
spacy==3.7.4
tiktoken==0.6.0